订单信息
git-svn-id: https://10.133.192.229/svn/CustomerManage/trunk/Code@2075 67192a19-bcbd-5648-baee-99b7ca3a6e28

+++ CPOS.BS.Entity/InoutInfo.cs
@@ -343,7 +343,7 @@ namespace JIT.CPOS.BS.Entity
         /// <summary>
         /// 行号
         /// </summary>
-        public int Row_No { get; set; }
+        public Int64 Row_No { get; set; }
         /// <summary>
         /// 总记录数
         /// </summary>
